Fix require changes
